fix test runner to catch failed commands
set -e does not work if the failed command is placed before a
&& or || operator. So now the test is executed in a subshell
where it is enabled the -e option. then you can check the exit
code instead of pipelining the command immediately.
the test class execution is placed inside a subshell as well
in order to automatically clean up sourced functions.